Вынесла константы в отдельный файл
