+++ js/constants/generation.js
@@ -1,0 +1,38 @@
+const NUMBER_ID_PHOTO = 25;
+const NUMBER_ID_COMMENT = 1000;
+
+const DESCRIPTIONS = [
+  'Превосходные детали!',
+  'Настоящее единение.',
+  'Что у него с лицом?',
+  'Лучше бы пошёл домой.',
+  'Лучший вид из окна сегодня.',
+  'Всегда знал, что чёрный мне к лицу.',
+  'Вклад в будущее, раз Тинькофф не разрешает.',
+  'Как неожиданно и приятно. Очень приятно!'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const USER_NAMES = [
+  'Андрей',
+  'Анна',
+  'Анастасия',
+  'Евгений',
+  'Федор',
+  'Максим',
+  'Татьяна',
+  'Сергей',
+  'Петр',
+  'Вячеслав',
+  'Марьяна'
+];
+
+export { NUMBER_ID_COMMENT, NUMBER_ID_PHOTO, DESCRIPTIONS, MESSAGES, USER_NAMES };
